+++ src/pages/Promotion/Search/Search.css
@@ -1,0 +1,4 @@
+.main {
+  max-width: 800px;
+  margin: 30px auto;
+}
